+++ src/main/java/com/ezezbiz/demo/files/excel/ReadExcelDemo.java
@@ -1,0 +1,58 @@
+package com.ezezbiz.demo.files.excel;
+
+import com.google.common.io.Resources;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.xssf.usermodel.XSSFSheet;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+
+import java.io.InputStream;
+import java.util.Iterator;
+
+public class ReadExcelDemo {
+    public static void main(String[] args)
+    {
+        try
+        {
+            InputStream file = Resources.getResource("SampleData.xlsx").openStream();
+
+            //Create Workbook instance holding reference to .xlsx file
+            XSSFWorkbook workbook = new XSSFWorkbook(file);
+
+            //Get first/desired sheet from the workbook
+            XSSFSheet sheet = workbook.getSheetAt(1);
+
+            //Iterate through each rows one by one
+            Iterator<Row> rowIterator = sheet.iterator();
+            while (rowIterator.hasNext())
+            {
+                Row row = rowIterator.next();
+                //For each row, iterate through all the columns
+                Iterator<Cell> cellIterator = row.cellIterator();
+
+                while (cellIterator.hasNext())
+                {
+                    Cell cell = cellIterator.next();
+                    System.out.print(cell.getCellType()+"\t");
+
+                    //Check the cell type and format accordingly
+                    switch (cell.getCellType())
+                    {
+                        case NUMERIC:
+                            System.out.print(cell.getNumericCellValue() + "\t");
+                            break;
+                        case STRING:
+                            System.out.print(cell.getStringCellValue() + "\t");
+                            break;
+                    }
+                }
+                System.out.println("");
+            }
+            file.close();
+        }
+        catch (Exception e)
+        {
+            e.printStackTrace();
+        }
+    }
+}
