+++ src/main/java/com/ezezbiz/demo/files/io/FileReadWriteBufferTestMultiThread.java
@@ -1,4 +1,4 @@
-package com.ezezbiz.demo.tuning.file.buffer;
+package com.ezezbiz.demo.files.io;
 
 import java.io.File;
 import java.io.FileInputStream;
